refactor(LottoTest) : 미션으로 제공된 LottoTest랑 새로만든거랑 결합
